feat: changes files (from js to tsx)
