+++ components/black-friday/BestStorageSection.tsx
@@ -7,17 +7,19 @@ const BestStorageSection = ({ textContent, lang }) => {
     <section className="overflow-hidden">
       <div className="flex flex-col items-center py-16 px-10 text-center md:px-20">
         <div className="top-24 flex flex-col items-center space-y-5">
-          <h1 className="text-4xl font-semibold text-white">{textContent.BestStorage.title}</h1>
-          <h3 className="text-xl text-white md:font-normal">{textContent.BestStorage.subtitle}</h3>
+          <h2 className="max-w-[526px] text-4xl font-semibold text-white md:text-5xl">
+            {textContent.BestStorage.title}
+          </h2>
+          <h3 className="text-xl text-white">{textContent.BestStorage.subtitle}</h3>
         </div>
         <div className="py-9">
           <ButtonDeal lang={lang} />
         </div>
         <img
           className="hidden md:flex lg:max-w-3xl"
-          src="/images/special-offer/black-friday/Devices.png"
+          src="/images/home/Internxt-secure-cloud-storage.webp"
+          alt="Internxt secure cloud storage"
           draggable="false"
-          alt="Desktop, laptop and phone with Internxt app"
         />
         <img
           className="flex md:hidden"
