+++ components/black-friday/FAQSection.tsx
