+++ components/black-friday/FeatureSection.tsx
