+++ components/black-friday/FooterSection.tsx
