+++ components/black-friday/PlatformSection.tsx
