+++ components/black-friday/SuiteSection.tsx
@@ -8,8 +8,8 @@ const SuiteSection = ({ textContent, lang }) => {
       <div className="z-10 flex flex-col items-center py-16">
         <div className="px-6 text-left sm:text-center">
           <div className="flex flex-col items-center px-5 pb-16 text-center">
-            <h1 className="text-4xl font-semibold">{textContent.SuiteSection.title}</h1>
-            <p className="max-w-[756px] pt-5 text-xl font-normal">{textContent.SuiteSection.subtitle}</p>
+            <h2 className="text-4xl font-semibold text-gray-100 md:text-5xl">{textContent.SuiteSection.title}</h2>
+            <p className="max-w-[756px] pt-5 text-xl text-gray-80">{textContent.SuiteSection.subtitle}</p>
           </div>
 
           <div className="flex flex-col space-y-20 text-left text-white lg:grid lg:grid-cols-1 lg:grid-rows-2 lg:gap-20 lg:space-y-0">
@@ -24,7 +24,7 @@ const SuiteSection = ({ textContent, lang }) => {
 
               <div className="lg:pl-15 relative mt-16 flex self-stretch  lg:mt-0">
                 <div className="hidden lg:flex lg:max-w-[480px]">
-                  <img src="/images/privacy/Internxt-Drive.webp" alt="Internxt Drive secure service" />
+                  <img src="/images/privacy/Internxt-Drive.webp" alt="Internxt Drive" draggable={false} />
                 </div>
               </div>
             </div>
@@ -40,7 +40,7 @@ const SuiteSection = ({ textContent, lang }) => {
 
               <div className="lg:pl-15 relative mt-16 flex self-stretch  lg:mt-0">
                 <div className="hidden lg:flex lg:max-w-[480px]">
-                  <img src="/images/privacy/photos-image.png" alt="Internxt Photos secure service" />
+                  <img src="/images/privacy/photos-image.png" alt="Internxt Photos" />
                 </div>
               </div>
             </div>
@@ -55,13 +55,7 @@ const SuiteSection = ({ textContent, lang }) => {
 
               <div className="lg:pl-15 relative mt-16 flex self-stretch lg:mt-0">
                 <div className="hidden lg:flex lg:max-w-[480px]">
-                  <img
-                    src="/images/privacy/Internxt-Send.png"
-                    width={480}
-                    height={480}
-                    draggable={false}
-                    alt="Internxt Send transfer service"
-                  />
+                  <img src="/images/privacy/Share-by-email.webp" alt="Internxt Send" />
                 </div>
               </div>
             </div>
