+++ components/layout/components/LanguageBox.tsx
@@ -24,7 +24,7 @@ const selectedLang = {
   ru: 'RU',
 };
 
-export default function LanguageBox({ darkMode }) {
+export default function LanguageBox({ darkMode, isBlackFriday }) {
   const router = useRouter();
 
   const [currentLangText, setCurrentLangText] = useState<string>(selectedLang[router.locale]);
@@ -50,72 +50,95 @@ export default function LanguageBox({ darkMode }) {
         <div className="absolute -top-4 left-1/2 h-4 w-4/5 -translate-x-1/2" />
 
         <div className="relative grid gap-0 whitespace-nowrap lg:grid-cols-1">
-          <Link href={router.pathname} locale="en">
-            <a
-              className={`flex flex-row justify-start rounded-lg py-2 px-4 text-base font-medium text-cool-gray-80  hover:bg-gray-1`}
-              onClick={() => changeLang('en')}
-            >
-              {currentLang.en}
-            </a>
-          </Link>
+          {isBlackFriday ? (
+            <>
+              <Link href={router.pathname} locale="en">
+                <a
+                  className={`flex flex-row justify-start rounded-lg py-2 px-4 text-base font-medium text-cool-gray-80  hover:bg-gray-1`}
+                  onClick={() => changeLang('en')}
+                >
+                  {currentLang.en}
+                </a>
+              </Link>
+              <Link href={router.pathname} locale="fr">
+                <a
+                  className={`flex flex-row justify-start rounded-lg py-2 px-4 text-base font-medium text-cool-gray-80  hover:bg-gray-1`}
+                  onClick={() => changeLang('fr')}
+                >
+                  {currentLang.fr}
+                </a>
+              </Link>
+            </>
+          ) : (
+            <>
+              <Link href={router.pathname} locale="en">
+                <a
+                  className={`flex flex-row justify-start rounded-lg py-2 px-4 text-base font-medium text-cool-gray-80  hover:bg-gray-1`}
+                  onClick={() => changeLang('en')}
+                >
+                  {currentLang.en}
+                </a>
+              </Link>
 
-          <Link href={router.pathname} locale="es">
-            <a
-              className={`flex flex-row justify-start rounded-lg py-2 px-4 text-base font-medium text-cool-gray-80  hover:bg-gray-1`}
-              onClick={() => changeLang('es')}
-            >
-              {currentLang.es}
-            </a>
-          </Link>
+              <Link href={router.pathname} locale="es">
+                <a
+                  className={`flex flex-row justify-start rounded-lg py-2 px-4 text-base font-medium text-cool-gray-80  hover:bg-gray-1`}
+                  onClick={() => changeLang('es')}
+                >
+                  {currentLang.es}
+                </a>
+              </Link>
 
-          <Link href={router.pathname} locale="fr">
-            <a
-              className={`flex flex-row justify-start rounded-lg py-2 px-4 text-base font-medium text-cool-gray-80  hover:bg-gray-1`}
-              onClick={() => changeLang('fr')}
-            >
-              {currentLang.fr}
-            </a>
-          </Link>
-          <Link href={router.pathname} locale="de">
-            <a
-              className={`flex flex-row justify-start rounded-lg py-2 px-4 text-base font-medium text-cool-gray-80  hover:bg-gray-1 ${
-                darkMode ? 'text-white' : 'text-cool-gray-80'
-              }`}
-              onClick={() => changeLang('de')}
-            >
-              {currentLang.de}
-            </a>
-          </Link>
-          <Link href={router.pathname} locale="it">
-            <a
-              className={`flex flex-row justify-start rounded-lg py-2 px-4 text-base font-medium text-cool-gray-80  hover:bg-gray-1 ${
-                darkMode ? 'text-white' : 'text-cool-gray-80'
-              }`}
-              onClick={() => changeLang('it')}
-            >
-              {currentLang.it}
-            </a>
-          </Link>
-          <Link href={router.pathname} locale="zh">
-            <a
-              className={`flex flex-row justify-start rounded-lg py-2 px-4 text-base font-medium text-cool-gray-80  hover:bg-gray-1 ${
-                darkMode ? 'text-white' : 'text-cool-gray-80'
-              }`}
-              onClick={() => changeLang('zh')}
-            >
-              {currentLang.zh}
-            </a>
-          </Link>
-          <Link href={router.pathname} locale="ru">
-            <a
-              className={`flex flex-row justify-start rounded-lg py-2 px-4 text-base font-medium text-cool-gray-80  hover:bg-gray-1 ${
-                darkMode ? 'text-white' : 'text-cool-gray-80'
-              }`}
-              onClick={() => changeLang('ru')}
-            >
-              {currentLang.ru}
-            </a>
-          </Link>
+              <Link href={router.pathname} locale="fr">
+                <a
+                  className={`flex flex-row justify-start rounded-lg py-2 px-4 text-base font-medium text-cool-gray-80  hover:bg-gray-1`}
+                  onClick={() => changeLang('fr')}
+                >
+                  {currentLang.fr}
+                </a>
+              </Link>
+              <Link href={router.pathname} locale="de">
+                <a
+                  className={`flex flex-row justify-start rounded-lg py-2 px-4 text-base font-medium text-cool-gray-80  hover:bg-gray-1 ${
+                    darkMode ? 'text-white' : 'text-cool-gray-80'
+                  }`}
+                  onClick={() => changeLang('de')}
+                >
+                  {currentLang.de}
+                </a>
+              </Link>
+              <Link href={router.pathname} locale="it">
+                <a
+                  className={`flex flex-row justify-start rounded-lg py-2 px-4 text-base font-medium text-cool-gray-80  hover:bg-gray-1 ${
+                    darkMode ? 'text-white' : 'text-cool-gray-80'
+                  }`}
+                  onClick={() => changeLang('it')}
+                >
+                  {currentLang.it}
+                </a>
+              </Link>
+              <Link href={router.pathname} locale="zh">
+                <a
+                  className={`flex flex-row justify-start rounded-lg py-2 px-4 text-base font-medium text-cool-gray-80  hover:bg-gray-1 ${
+                    darkMode ? 'text-white' : 'text-cool-gray-80'
+                  }`}
+                  onClick={() => changeLang('zh')}
+                >
+                  {currentLang.zh}
+                </a>
+              </Link>
+              <Link href={router.pathname} locale="ru">
+                <a
+                  className={`flex flex-row justify-start rounded-lg py-2 px-4 text-base font-medium text-cool-gray-80  hover:bg-gray-1 ${
+                    darkMode ? 'text-white' : 'text-cool-gray-80'
+                  }`}
+                  onClick={() => changeLang('ru')}
+                >
+                  {currentLang.ru}
+                </a>
+              </Link>
+            </>
+          )}
         </div>
       </div>
     </div>
